requirements.txt
python-binance
python-dotenv
flask
flask-cors